BACKEND/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-bcrypt
python-dotenv

pymysql
cryptography


yfinance
pandas
numpy


ta

xgboost
scikit-learn
joblib

vaderSentiment

finnhub-python

requests


upstash-redis

flask-jwt-extended
flask-bcrypt
pymysql
cryptography